feat(live2d): indicate when a story ends
