+++ src/pages/storyreader-live2d/StoryReaderLive2DStage.tsx
@@ -1,0 +1,55 @@
+import { useRef, forwardRef, useImperativeHandle, useEffect } from "react";
+
+import { useApp } from "@pixi/react";
+
+import { extensions, TickerPlugin } from "pixi.js";
+extensions.add(TickerPlugin);
+
+import { Live2DController } from "../../utils/Live2DPlayer/Live2DController";
+import { LoadStatus } from "../../utils/Live2DPlayer/types.d";
+import type { ILive2DControllerData } from "../../utils/Live2DPlayer/types.d";
+
+const StoryReaderLive2DStage = forwardRef<
+  { controller: Live2DController; reloadStage: () => void },
+  {
+    stageSize: [number, number];
+    controllerData: ILive2DControllerData;
+    onModelLoad: (status: LoadStatus) => void;
+  }
+>(({ stageSize, controllerData, onModelLoad }, ref) => {
+  const app = useApp();
+  const controller = useRef<Live2DController>();
+  useImperativeHandle(ref, () => {
+    return {
+      controller: controller.current!,
+      reloadStage: reloadStage,
+    };
+  });
+  useEffect(() => {
+    controller.current?.set_stage_size(stageSize);
+  }, [stageSize]);
+  useEffect(() => {
+    controller.current = new Live2DController(app, stageSize, controllerData);
+    //DEBUG
+    //window.controller = controller.current;
+    //DEBUG/
+    if (controller.current.layers.live2d.load_status() === "ready") {
+      onModelLoad(LoadStatus.Loading);
+      controller.current.layers.live2d.clear();
+      controller.current.live2d_load_model(0).then(() => {
+        onModelLoad(LoadStatus.Loaded);
+      });
+    }
+    return () => {
+      controller.current?.unload();
+    };
+  }, []);
+  function reloadStage() {
+    controller.current = new Live2DController(app, stageSize, controllerData);
+    controller.current.layers.live2d.clear();
+    controller.current.live2d_load_model(0);
+  }
+  return null;
+});
+StoryReaderLive2DStage.displayName = "StoryReaderLive2DStage";
+export default StoryReaderLive2DStage;
